refactor: move scalar prefilter to module
